feat(notification): implement permission handling
- Add runtime notification permission request for Android 13+
- Add toast messages for permission status feedback
- Create CompositionLocal for NotificationService
- Fix notification service implementation
- Update manifest with required permissions
- Add error handling for notification display

This change ensures proper notification permission handling
and provides clear feedback to users about permission status.
It also improves the reliability of notification delivery
through better error handling and proper Android 13+ support.

+++ app/src/main/java/com/sios/tech/plantreminderapp/notification/WateringReminderReceiver.kt
@@ -33,7 +33,7 @@ class WateringReminderReceiver : BroadcastReceiver() {
         val plantName = intent.getStringExtra(EXTRA_PLANT_NAME) ?: return
 
         if (plantId != -1L) {
-            notificationService.showWateringNotification(plantId, plantName)
+            notificationService.showWateringNotification(plantId.toInt(), plantName)
         }
     }
